+++ FsChessPgn/Util.fs
@@ -10,9 +10,9 @@ module Util =
     let Plyr i = enum<Player> (i)
     let BitB i =  Microsoft.FSharp.Core.LanguagePrimitives.EnumOfValue<uint64,Bitboard> (i)
 
-    let (-!) (r:Rank) (i:int):Rank = (int(r)-i)
-    let (+!) (r:Rank) (i:int):Rank = (int(r)+i)
+    let (-!) (r:Rank) (i:int16):Rank = r-i
+    let (+!) (r:Rank) (i:int16):Rank = r+i
 
-    let (--) (f:File) (i:int):File = (int(f)-i)
-    let (++) (f:File) (i:int):File = (int(f)+i)
+    let (--) (f:File) (i:int16):File = f-i
+    let (++) (f:File) (i:int16):File = f+i
